Add js function to format last sensor readings data, Add real tray name, crop type and last reading in the front end via js, Add real target settings to humidity, temperature and sun light to the front end via js

+++ SmartTray/SmartTray.Domain/DTO/TraySensorReadingDTO.cs
@@ -3,9 +3,9 @@
     public class TraySensorReadingDTO
     {
         public int TargetLightMinutes { get; set; }
-        public double DailyLightMinutes { get; set; }
+        public int DailyLightMinutes { get; set; }
         public int ArtificialLightMinutes { get; set; }
         public int SolarLightMinutes { get; set; }
-        public double RemainingLightMinutes { get; set; }
+        public int RemainingLightMinutes { get; set; }
     }
 }
